#12853 remove Environment.put() static from API, add Environment.ensure(name, ClassLoader) instead and make Environment.Named package-protected
