+++ jetty-core/jetty-util/src/main/java/org/eclipse/jetty/util/component/Environment.java
@@ -13,14 +13,10 @@
 
 package org.eclipse.jetty.util.component;
 
-import java.io.IOException;
 import java.util.Collection;
 import java.util.Collections;
-import java.util.Map;
-import java.util.concurrent.ConcurrentSkipListMap;
 
 import org.eclipse.jetty.util.Attributes;
-import org.eclipse.jetty.util.TypeUtil;
 
 /**
  * A named runtime environment containing a {@link ClassLoader} and {@link Attributes}.
@@ -32,22 +28,22 @@ public interface Environment extends Attributes
 
     static Collection<Environment> getAll()
     {
-        return Collections.unmodifiableCollection(Named.__environments.values());
+        return Collections.unmodifiableCollection(NamedEnvironment.__environments.values());
     }
     
     static Environment get(String name)
     {
-        return Named.__environments.get(name);
+        return NamedEnvironment.__environments.get(name);
     }
 
     static Environment ensure(String name)
     {
-        return Named.__environments.computeIfAbsent(name, Named::new);
+        return ensure(name, null);
     }
 
-    static Environment set(Environment environment)
+    static Environment ensure(String name, ClassLoader classLoader)
     {
-        return Named.__environments.put(environment.getName(), environment);
+        return NamedEnvironment.__environments.computeIfAbsent(name, n -> new NamedEnvironment(n, classLoader));
     }
 
     /**
@@ -56,7 +52,7 @@ public interface Environment extends Attributes
     String getName();
 
     /**
-     * @return The {@link ClassLoader} for the environment or if non set, then the {@link ClassLoader} that
+     * @return The {@link ClassLoader} for the environment or if none set, then the {@link ClassLoader} that
      * loaded the environment implementation.
      */
     ClassLoader getClassLoader();
@@ -79,56 +75,4 @@ public interface Environment extends Attributes
             Thread.currentThread().setContextClassLoader(old);
         }
     }
-
-    class Named extends Attributes.Mapped implements Environment, Dumpable
-    {
-        private static final Map<String, Environment> __environments = new ConcurrentSkipListMap<>(String.CASE_INSENSITIVE_ORDER);
-        private final String _name;
-        private final ClassLoader _classLoader;
-
-        public Named(String name)
-        {
-            this(name, null);
-        }
-
-        public Named(String name, ClassLoader classLoader)
-        {
-            _name = name;
-            _classLoader = classLoader == null ? this.getClass().getClassLoader() : classLoader;
-        }
-
-        @Override
-        public String getName()
-        {
-            return _name;
-        }
-
-        @Override
-        public ClassLoader getClassLoader()
-        {
-            return _classLoader;
-        }
-
-        @Override
-        public String dump()
-        {
-            return Dumpable.dump(this);
-        }
-
-        @Override
-        public void dump(Appendable out, String indent) throws IOException
-        {
-            Dumpable.dumpObjects(out, indent,
-                this,
-                new ClassLoaderDump(getClassLoader()),
-                new DumpableCollection("Attributes " + _name, asAttributeMap().entrySet()));
-        }
-
-        @Override
-        public String toString()
-        {
-            return "%s@%x{%s}".formatted(TypeUtil.toShortName(this.getClass()), hashCode(), _name);
-        }
-    }
-
 }
